+++ src/components/birdrecord/BirdRecordKernel.java
@@ -1,3 +1,5 @@
+package components.birdrecord;
+
 import components.standard.Standard;
 
 /**