+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.02.04
+
+### Added
+
+- Designed a Bird component
+- Designed a Playable Character component
+- Designed a Timeline component
+
+## 2024.03.24
+
+### Added
+
+- Designed a proof of concept for Bird Record/component
+
 ## 2025.03.30
 
 ### Added
@@ -21,4 +35,25 @@ the following form: YYYY.0M.0D.
 ### Added
 
 - Designed abstract class for BirdRecord component
+
+## 2025.04.18
+
+### Added
+
+- Designed kernel implementation for BirdRecord component
+
+### Updated
+
+- Updated BirdRecordKernel.java with Standard methods transferFrom, clear, and newInstance.
+
+## 2025.04.18
+
+### Added
+
+- Designed test suite for BirdRecord component
+- Designed two different use cases for BirdRecord component
+
+### Updated
+
+- Changed design to polish up a lot of things.
 ```